% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_elp.R
\name{overwrite_elp}
\alias{overwrite_elp}
\alias{softwrite_elp}
\title{Over/soft write some default values to the ellipsis \code{...}.}
\usage{
overwrite_elp(..., append = NULL, warn = TRUE)

softwrite_elp(..., append = NULL, warn = FALSE)
}
\arguments{
\item{...}{the ellipsis \code{...} and default values to be setted.}

\item{append}{a list of arguments.
Mostly the return of \code{overwrite_elp} or \code{softwrite_elp}.}

\item{warn}{if \code{TRUE}, warn when conflicts user's specific and default value.}
}
\value{
overwrited ellipsis.
}
\description{
Set default value to the ellipsis \code{...}.
When conflicts user's specific and the default value,
\code{overwrite_elp()} prioritize default value
and \code{softwrite_elp()} user's specific.
Note that when chain over/soft write_elp,
use \code{overwrite_elp()} firstly or set \code{warn = FALSE} like example code..
}
\details{
The ellipsis \code{...} is used in for example the \code{plot} function frequently
to provide flexible visualization.
When we implement the \code{plot} function as the S3 generic method or others,
we want to set the default graphics parameters such as \code{xlab} or \code{title}
to \code{...} in some cases.
\code{overwrite_elp()} and \code{softwrite_elp()} can set defaults to \code{...} easily,
and control rejecting or accepting user's specific parameters.
}
\examples{
# plot function with xlim = c(0, 2 * pi) and col = 'red'
# but col is arbitrary changeable
plot2pi <- function(x, ...){
  elp <- overwrite_elp(..., x = x, xlim = c(0, 2 * pi), ylab = "amplitude")
  elp <- softwrite_elp(..., append = elp, col = "red")
  do.call(plot, elp)
}

plot2pi(sin)

# col and title is accepted
plot2pi(cos, col = "blue", main = "cosine curve")

# xlim and ylab is rejected and warned
plot2pi(tan, xlim = c(-pi, pi), xlab = "phase", ylab = "tangent")
}
